+++ package.json
@@ -56,7 +56,8 @@
     "typescript": "^4.9.5",
     "usehooks-ts": "^2.9.1",
     "uuid": "^9.0.1",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^2.1.4",
+    "event-source-polyfill": "^1.0.31"
   },
   "scripts": {
     "start": "set PORT=3001 && react-scripts start",
